code/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0

# Web dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Serial communication
pyserial>=3.5

# Data processing
scipy>=1.11.0

# Optional: For advanced visualizations
seaborn>=0.12.0
openpyxl>=3.1.0  # For Excel export

# Development tools
jupyter>=1.0.0
ipython>=8.0.0